pyproject.toml
[project]
name = "unsplash-client"
version = "0.1.0"
description = "Python client for the Unsplash API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "ruff>=0.13.1",
    "pre-commit>=4.3.0",
    "twine>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = [".venv", "dist", "build"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL", "RUF", "C90"]
ignore = [
    "E501",     # line too long
    "PLR2004",  # magic values
    "PLR0913",  # too many arguments
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88